fix!: Load the DATA_DIR via derived_settings tooling.
If we don't make the DATA_DIR a derived_setting, the only way to
override it is to override the entirety of the MODULESTORE config, or
override DATA_DIR in common.py in your fork.

BREAKING CHANGE: If you override the `MODULESTORE` config and you don't
have two stores defined under MixedModulestore, both whin an
['OPTIONS']['fs_root'] entry.  The service will fail to startup.  You
can work around this issue by adding a dummy entry for `fs_root` your
secondary store.