fix: Add an `fs_root` entry to bok_choy.yml configs.
In the long run we want to get rid of these configs but I don't want to
couple that with the changing of the migration jobs to a more minimal
config.  We need to add this because common.py now expects both
modulestore stores to have an `fs_root` entry that is treated as a
derived setting.  If you don't have an `fs_Root` `OPTIONS` entry, it is
now throwing an error.  We make this fix because the file is still being
used in the static asset check jobs so we can't get rid of it yet.